requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.2.0
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
environs==11.0.0
frozenlist==1.4.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
marshmallow==3.22.0
multidict==6.1.0
packaging==24.1
propcache==0.2.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.32.3
sendgrid==6.11.0
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
twilio==9.3.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
yarl==1.15.0
